# Conflicts: #	shop-admin/src/main/java/com/masiis/shop/admin/service/order/BOrderPayService.java #	shop-admin/src/main/java/com/masiis/shop/admin/service/order/BUpgradePayService.java #	shop-admin/src/main/java/com/masiis/shop/admin/service/promotion/SfUserTurnTableService.java #	shop-admin/src/main/java/com/masiis/shop/admin/service/user/PfUserRecommendRelationService.java #	shop-admin/src/main/resources/jdbc.properties #	shop-admin/src/main/webapp/WEB-INF/pages/delivery/bOrderList.jsp #	shop-dao/src/main/java/com/masiis/shop/dao/beans/order/BOrderAdd.java #	shop-dao/src/main/java/com/masiis/shop/dao/mall/order/SfOrderMapper.java #	shop-dao/src/main/java/com/masiis/shop/dao/mall/promotion/SfTurnTableRuleMapper.java #	shop-dao/src/main/java/com/masiis/shop/dao/mall/promotion/SfUserTurnTableMapper.java #	shop-dao/src/main/java/com/masiis/shop/dao/platform/order/PfBorderMapper.java #	shop-dao/src/main/java/com/masiis/shop/dao/platform/user/PfUserAgentApplicationMapper.java #	shop-dao/src/main/java/com/masiis/shop/dao/po/SfTurnTable.java #	shop-dao/src/main/resources/mapper/platform/order/PfBorderMapper.xml #	shop-dao/src/main/resources/mapper/platform/user/PfUserSkuMapper.xml #	shop-scheduler/src/main/webapp/index.jsp #	shop-service/src/main/java/com/masiis/shop/web/mall/service/order/SfOrderService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/order/BOrderAddService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/order/BOrderPayEndMessageService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/order/BOrderPayService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/order/BUpgradePayService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/order/OrderQueueTimeDealService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/product/PfUserAgentApplicationService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/user/PfUserRecommendRelationService.java #	shop-service/src/main/java/com/masiis/shop/web/platform/service/user/UpgradeWechatNewsService.java #	shop-service/src/main/java/com/masiis/shop/web/promotion/cpromotion/service/gorder/SfTurnTableRuleService.java #	shop-service/src/main/java/com/masiis/shop/web/promotion/cpromotion/service/guser/SfUserTurnTableService.java #	shop-web/src/main/java/com/masiis/shop/web/platform/controller/order/BOrderAddController.java #	shop-web/src/main/java/com/masiis/shop/web/platform/controller/order/BOrderController.java #	shop-web/src/main/java/com/masiis/shop/web/platform/controller/user/DevelopingController.java #	shop-web/src/main/java/com/masiis/shop/web/platform/controller/user/MyRecommendController.java #	shop-web/src/main/java/com/masiis/shop/web/platform/filter/LoginFilter.java

+++ shop-dao/src/main/java/com/masiis/shop/dao/po/PfUserStatistics.java
@@ -22,7 +22,7 @@ public class PfUserStatistics {
      */
     private Long userId;
 
-    private Long skuId;
+    private Integer skuId;
 
     /**
      * 总销售额
@@ -172,11 +172,11 @@ public class PfUserStatistics {
         this.remark = remark == null ? null : remark.trim();
     }
 
-    public Long getSkuId() {
+    public Integer getSkuId() {
         return skuId;
     }
 
-    public void setSkuId(Long skuId) {
+    public void setSkuId(Integer skuId) {
         this.skuId = skuId;
     }
